Add enhanced performance testing framework
- Introduced a new testing framework for enhanced performance scenarios, including configurations for action calls, payload management, and memory tracking.
- Implemented multiple test cases to evaluate performance under various conditions, including simple and complex scenarios with multiple brokers and services.
- Added comprehensive logging and metrics collection to facilitate analysis of test results and memory usage.
- Created new types and structures to support enhanced test configurations and results, improving the overall testing strategy for the project.

+++ test/performance/stress_test.go
@@ -59,7 +59,7 @@ type StressTestResult struct {
 	ErrorRate              float64
 	PeakMemoryUsage        uint64
 	FinalMemoryUsage       uint64
-	MemoryGrowth           uint64
+	MemoryGrowth           int64
 	PeakGoroutineCount     int
 	FinalGoroutineCount    int
 	GoroutineLeak          int
@@ -403,7 +403,7 @@ func runStressTestInternal(t *testing.T, config *StressTestConfig) *StressTestRe
 		ErrorRate:              errorRate,
 		PeakMemoryUsage:        peakMemory,
 		FinalMemoryUsage:       finalMem.HeapAlloc,
-		MemoryGrowth:           finalMem.HeapAlloc - initialMem.HeapAlloc,
+		MemoryGrowth:           int64(finalMem.HeapAlloc) - int64(initialMem.HeapAlloc),
 		PeakGoroutineCount:     finalGoroutines,
 		FinalGoroutineCount:    finalGoroutines,
 		GoroutineLeak:          finalGoroutines - initialGoroutines,
